+++ src/documentclient.ts
@@ -5,7 +5,7 @@ import * as tunnel from "tunnel";
 import * as url from "url";
 import * as util from "util";
 import { Base, ResponseCallback } from "./base";
-import { Constants, Helper, Platform } from "./common";
+import { Constants, Helper, Platform, StatusCodes, SubStatusCodes } from "./common";
 import { DocumentClientBase } from "./DocumentClientBase";
 import {
     ConnectionPolicy, ConsistencyLevel, DatabaseAccount, Document, PartitionKey, QueryCompatibilityMode,
@@ -13,32 +13,9 @@ import {
 import { GlobalEndpointManager } from "./globalEndpointManager";
 import { FetchFunctionCallback, IHeaders, SqlQuerySpec } from "./queryExecutionContext";
 import { QueryIterator } from "./queryIterator";
-import { RequestHandler, Response } from "./request";
+import { ErrorResponse, RequestHandler, Response } from "./request";
 import { RetryOptions } from "./retry";
 import { SessionContainer } from "./sessionContainer";
-
-// var Base = require("./base")
-//     , https = require("https")
-//     , url = require("url")
-//     , tunnel = require("tunnel")
-//     , AzureDocuments = require("./documents")
-//     , QueryIterator = require("./queryIterator")
-//     , RequestHandler = require("./request")
-//     , RetryOptions = require("./retryOptions")
-//     , GlobalEndpointManager = require("./globalEndpointManager")
-//     , Constants = require("./constants")
-//     , Helper = require("./helper").Helper
-//     , util = require("util")
-//     , Platform = require("./platform")
-//     , SessionContainer = require("./sessionContainer");
-
-// if (typeof exports !== "undefined") {
-//     exports.DocumentClient = DocumentClient;
-//     exports.DocumentBase = AzureDocuments;
-//     exports.RetryOptions = RetryOptions;
-//     exports.Base = Base;
-//     exports.Constants = Constants;
-// }
 
 export class DocumentClient extends DocumentClientBase {
     /**
@@ -999,7 +976,7 @@ export class DocumentClient extends DocumentClientBase {
                 this.applySessionToken(path, reqHeaders);
 
                 const { result, headers: resHeaders } = await documentclient.get(readEndpoint, request, reqHeaders);
-                this.captureSessionToken(path, Constants.OperationTypes.Query, reqHeaders, resHeaders);
+                this.captureSessionToken(undefined, path, Constants.OperationTypes.Query, resHeaders);
                 return this.processQueryFeedResponse({ result, headers: resHeaders }, !!query, resultFn, createFn);
             } else {
                 initialHeaders[Constants.HttpHeaders.IsQuery] = "true";
@@ -1024,7 +1001,7 @@ export class DocumentClient extends DocumentClientBase {
                 const response =
                     await documentclient.post(readEndpoint, request, query, reqHeaders);
                 const { result, headers: resHeaders } = response;
-                this.captureSessionToken(path, Constants.OperationTypes.Query, reqHeaders, resHeaders);
+                this.captureSessionToken(undefined, path, Constants.OperationTypes.Query, resHeaders);
                 return this.processQueryFeedResponse({ result, headers: resHeaders }, !!query, resultFn, createFn);
             }
 
@@ -2372,9 +2349,10 @@ export class DocumentClient extends DocumentClientBase {
 
             const writeEndpoint = await this._globalEndpointManager.getWriteEndpoint();
             const { result, headers: resHeaders } = await this.post(writeEndpoint, path, body, requestHeaders);
-            this.captureSessionToken(path, Constants.OperationTypes.Create, requestHeaders, resHeaders);
+            this.captureSessionToken(undefined, path, Constants.OperationTypes.Create, resHeaders);
             return Base.ResponseOrCallback(callback, { result, headers: resHeaders });
         } catch (err) {
+            this.captureSessionToken(err, path, Constants.OperationTypes.Upsert, (err as ErrorResponse).headers);
             Base.ThrowOrCallback(callback, err);
         }
     }
@@ -2394,9 +2372,10 @@ export class DocumentClient extends DocumentClientBase {
             // upsert will use WriteEndpoint since it uses POST operation
             const writeEndpoint = await this._globalEndpointManager.getWriteEndpoint();
             const { result, headers: resHeaders } = await this.post(writeEndpoint, path, body, requestHeaders);
-            this.captureSessionToken(path, Constants.OperationTypes.Upsert, requestHeaders, resHeaders);
+            this.captureSessionToken(undefined, path, Constants.OperationTypes.Upsert, resHeaders);
             return Base.ResponseOrCallback(callback, { result, headers: resHeaders });
         } catch (err) {
+            this.captureSessionToken(err, path, Constants.OperationTypes.Upsert, (err as ErrorResponse).headers);
             Base.ThrowOrCallback(callback, err);
         }
     }
@@ -2419,9 +2398,10 @@ export class DocumentClient extends DocumentClientBase {
             // replace will use WriteEndpoint since it uses PUT operation
             const writeEndpoint = await this._globalEndpointManager.getWriteEndpoint();
             const result = await this.put(writeEndpoint, path, resource, reqHeaders);
-            this.captureSessionToken(path, Constants.OperationTypes.Replace, reqHeaders, result.headers);
+            this.captureSessionToken(undefined, path, Constants.OperationTypes.Replace, result.headers);
             return Base.ResponseOrCallback(callback, result);
         } catch (err) {
+            this.captureSessionToken(err, path, Constants.OperationTypes.Upsert, (err as ErrorResponse).headers);
             Base.ThrowOrCallback(callback, err);
         }
     }
@@ -2449,9 +2429,10 @@ export class DocumentClient extends DocumentClientBase {
             // read will use ReadEndpoint since it uses GET operation
             const readEndpoint = await this._globalEndpointManager.getReadEndpoint();
             const response = await this.get(readEndpoint, request, requestHeaders);
-            this.captureSessionToken(path, Constants.OperationTypes.Read, requestHeaders, response.headers);
+            this.captureSessionToken(undefined, path, Constants.OperationTypes.Read, response.headers);
             return response;
         } catch (err) {
+            this.captureSessionToken(err, path, Constants.OperationTypes.Upsert, (err as ErrorResponse).headers);
             throw err;
         }
     }
@@ -2474,13 +2455,14 @@ export class DocumentClient extends DocumentClientBase {
             const writeEndpoint = await this._globalEndpointManager.getWriteEndpoint();
             const response = await this.delete(writeEndpoint, path, reqHeaders);
             if (Base.parseLink(path).type !== "colls") {
-                this.captureSessionToken(path, Constants.OperationTypes.Delete, reqHeaders, response.headers);
+                this.captureSessionToken(undefined, path, Constants.OperationTypes.Delete, response.headers);
             } else {
                 this.clearSessionToken(path);
             }
             return Base.ResponseOrCallback(callback, response);
 
         } catch (err) {
+            this.captureSessionToken(err, path, Constants.OperationTypes.Upsert, (err as ErrorResponse).headers);
             Base.ThrowOrCallback(callback, err);
         }
     }
@@ -2653,7 +2635,7 @@ export class DocumentClient extends DocumentClientBase {
                 `The "headers" parameter must be an instance of "Object". Actual type is: "${typeof headers}".`);
         }
 
-        headers[Constants.HttpHeaders.IsUpsert] = true;
+        (headers as IHeaders)[Constants.HttpHeaders.IsUpsert] = true;
     }
 
     /**
@@ -2697,10 +2679,15 @@ export class DocumentClient extends DocumentClientBase {
         }
     }
 
-    public captureSessionToken(path: string, opType: string, reqHeaders: IHeaders, resHeaders: IHeaders) {
+    public captureSessionToken(err: ErrorResponse, path: string, opType: string, resHeaders: IHeaders) {
         const request: any = this.getSessionParams(path); // TODO: any request
         request.operationType = opType;
-        this.sessionContainer.setSessionToken(request, reqHeaders, resHeaders);
+        if (!err ||
+            ((!this.isMasterResource(request.resourceType)) &&
+                (err.code === StatusCodes.PreconditionFailed || err.code === StatusCodes.Conflict ||
+                    (err.code === StatusCodes.NotFound && err.substatus !== SubStatusCodes.ReadSessionNotAvailable)))) {
+            this.sessionContainer.setSessionToken(request, resHeaders);
+        }
     }
 
     public clearSessionToken(path: string) {
@@ -2726,6 +2713,21 @@ export class DocumentClient extends DocumentClientBase {
             resourceAddress,
             resourceType,
         };
+    }
+
+    public isMasterResource(resourceType: string): boolean {
+        if (resourceType === Constants.Path.OffersPathSegment ||
+            resourceType === Constants.Path.DatabasesPathSegment ||
+            resourceType === Constants.Path.UsersPathSegment ||
+            resourceType === Constants.Path.PermissionsPathSegment ||
+            resourceType === Constants.Path.TopologyPathSegment ||
+            resourceType === Constants.Path.DatabaseAccountPathSegment ||
+            resourceType === Constants.Path.PartitionKeyRangesPathSegment ||
+            resourceType === Constants.Path.CollectionsPathSegment) {
+            return true;
+        }
+
+        return false;
     }
 }
 
@@ -2805,6 +2807,7 @@ export interface RequestOptions {
  * A value indicating whether users are enabled to send more than one request to execute the query \
  * in the Azure Cosmos DB database service.
  * <p>More than one request is necessary if the query is not scoped to single partition key value.</p>
+ * @property {boolean} [populateQueryMetrics]       -       Whether to populate the query metrics.
  * @property {boolean} [enableScanInQuery]          -       \
  * Allow scan on the queries which couldn't be served as indexing was opted out on the requested paths.
  * @property {number} [maxDegreeOfParallelism]      -       \
@@ -2821,6 +2824,7 @@ export interface FeedOptions {
     continuation?: string;
     disableRUPerMinuteUsage?: boolean;
     enableCrossPartitionQuery?: boolean;
+    populateQueryMetrics?: boolean;
     enableScanInQuery?: boolean;
     maxDegreeOfParallelism?: number;
     maxItemCount?: number;